Add rest api --port option
Previously, BisqCommandLineTest would fail if bisqd was already running
on the machine due to the default port 2140 already being bound. This
change allows the user (and integration tests) to provide a --port
option to assign a different port, where a value of 0 will assign a
random available port. The net effect is that tests now pass as expected
regardless of whether a daemon is already bound to 2140.

This was a useful excercise in understanding how to inherit and reuse
options from top level commands in subcommands within the picocli
programming model.

It is, in any case, a naive approach to customizing the api endpoint.
Later commits will refine it, and the --port api option will probably
not survive long as-is.

Note that the way this change works with @ParentCommand is not ideal and
an improvement has been suggested at remkop/picocli#1385.

Fixes #11