build: Enforce and auto-provision JDK 16
Previously the build would fail on compilation with detailed syntax
errors if the user's JDK was less than version 16.

This change uses Gradle's relatively new java toolchains support to
specify that the required java / jdk version is 16. This support
includes attempt to provision such a JDK if it is not already resident
on the user's machine.

Caveat: this auto-provisioning will not work on M1 Macs at time of
writing however, and users should install the Azul JDK 16 manually,
which will then be automatically recognized and used by Gradle's
toolchain support.

An advantage of the toolchain support is that it allows Gradle itself to
run on one java version, say JDK 11, and to switch to JDK 16 when
building user code. This means that builds are more likely to "just
work" without having to update the active JDK in the user's environment.

An apparent downside of the toolchain approach is that it affords no way
to specify a _minimum_ version, e.g. JDK 16 or better. If 16 is
specified, that exact JDK version must be provisioned on the user's
machine, even if, say, JDK 17 is already there.

See:
 - ./gradlew -q javaToolChains
 - https://docs.gradle.org/current/userguide/toolchains.html

Fixes #12