fx: List offers
BisqFX now renders the current set of offers in the remote Bisq daemon,
and keeps this list up to date by subscribing to a WebSocket service
that publishes offer creation and deletion events.

The net effect is that the BisqFX offer list is updated in real-time
without requiring polling.

To try this out:

    $ bisqfx
    ... gui appears with empty offer list ...

    $ bisq offer create foo
    $ bisq offer create bar
    $ bisq offer create baz

Notice that the gui now contains all three offers.

    $ bisq offer delete all

Notice that the gui's offer list is now empty

Now try starting a second bisqfx and repeating the creation / deletion
events. Notice that both guis remain updated in real time.

Fixes: #19