Extract daemon subproject from core
In order to clearly separate the daemon application from the underlying
bisq node that it starts. Previously the core subproject was acting both
as a library and as an application. Now the daemon subproject is the
application and core is the library. This allows daemon to have its own
logback.xml and avoid having multiple such files on the classpath in the
case of the fx app. Fx and cli (tests) depend on core in order to be
able to spin up a node, but they do not need the main method of the
daemon itself.

Fixes #21