Introduce `bisq node` and .bisq/config file
A `bisq node` subcommand is now supported that reports the set of all
known bisq nodes as configured in a new ~/.bisq/config file. The
currently selected node is marked with an asterisk and highlighted in
green similar to the way `git branch` behaves. Fixes #24.

Configuration values can come from three sources: defaults as set in the
command classes, options passed on the command line, and options
declared in the config file. This change ensures that all options are
detected and selected with the correct precedence and priority over one
another. Command line is the highest priority, config file is secondary,
and defaults are the fallback. Fixes #25.

This commit also introduces an explicit --help option and customizes the
color scheme of the resulting usage text to be Bisq-themed. Fixes #26.

The 'truth' assertion library has also been introduced to make testing
easier. BisqCLITest has been refactored to be more robust in the face of
changes that affect usage text output.

This change also implements a novel approach to initializing picocli
commands prior to execution, see remkop/picocli#1385 for details.