fx: Shut down resources on JavaFX close request
Previously, bisqfx required a kill signal in order to shut down because
various background processes were not being stopped when the user issued
a request to close the application.

Now, BisqApiClient and OfferBookApiClient implement AutoCloseable and
their close() methods are called by the JavaFX onCloseRequest hook. This
ensures that websocket connections, etc are stopped, allowing the
application to exit as expected.