fx: Add help, version, node and verbose options
Prior to this change, `bisqfx` was hardcoded to query a local node. Like
`bisq`, it can now accept a --node option to allow pointing it at a node
running on a different machine or a non-default port.

This required introducing picocli to bisqfx as has already been done in
the cli subproject. To get reuse between them, a fair bit of common
infrastrucutre has been extracte cli up into the api subproject.

This is just a quick implementation for demo purposes, it does not yet
look to the bisq config file for node definitions, in the same way
`bisq` does, etc. This means that when invoking `bisqfx` it is necessary
to use full hostnames as the argument to the `--node` option, e.g.:

    $ bisqfx --node myhost.com:2140

bisqfx-completion.bash is now generated and the readme has been updated
to reflect.

Now that there is a --verbose option, the default log level for bisqfx
has been changed from DEBUG to INFO. to set it back to DEBUG via the
command line, run:

    $ bisqfx -v [...]

Fixes #30