docs: Create glossary
The intent here is to establish and refine a Ubiquitous Language (in the
DDD sense) that can help guide our efforts in building out the ambitious
feature set of Bisq 2.0. The terms laid out here attempt to cover most
of the points mentioned in the original Bisq 2.0 proposal at
bisq-network/proposals#330 and also include some original insights that
emerged during the process. The document is and will remain a work in
progress, but is now in good enough shape to share and get feedback on.

Fixes #31